// frontend/src/components/chatHistoryPanel/chatHistoryPanel.module.scss.d.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

declare const styles: {
  leftPanelContainer: string;
  renderNewChat: string;
  searchInput: string;
  listItem: string;
  listItemHover: string;
  iconButton: string;
  newChat: string;
  searchContainer: string;
  chatList: string;
  chatOptions: string;
  chatListDetails: string;
  chatListInfo: string;
  editOption: string;
  deleteChat: string;
  hoverChatBackground: string;
  transparentChatBackground: string;
  editChat: string;
  chatHistoryTimestamp: string;
  alterChatButtons: string;
  activeChat: string;
};

export default styles;